requirements.txt
# core libs
langchain==0.3.25          # orchestrator
langchain-community==0.3.24
langchain-text-splitters==0.3.8
langchain-openai==0.3.16

# vector store
chromadb==0.4.24

# utilities
python-dotenv==1.0.0
tiktoken~=0.9
pydantic>=2.7,<3
pandas==2.2.3
pdfplumber==0.11.6
PyPDF2==3.0.1
fastapi==0.111.0
uvicorn[standard]==0.29.0
beautifulsoup4 
selenium>=4.0.0
webdriver-manager>=3.0.0
requests
pytest